+++ driver-core/src/main/java/com/datastax/driver/core/Connection.java
@@ -16,7 +16,6 @@
 package com.datastax.driver.core;
 
 import javax.net.ssl.SSLEngine;
-import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.util.Iterator;
 import java.util.concurrent.*;
@@ -382,7 +381,7 @@ class Connection {
                     } else {
                         ce = new TransportException(address, "Error writing", writeFuture.getCause());
                     }
-                    handler.callback.onException(Connection.this, defunct(ce), System.nanoTime() - handler.startTime);
+                    handler.callback.onException(Connection.this, defunct(ce), System.nanoTime() - handler.startTime, handler.retryCount);
                 } else {
                     logger.trace("{} request sent successfully", Connection.this);
                 }
@@ -642,7 +641,7 @@ class Connection {
                     return;
                 }
                 handler.cancelTimeout();
-                handler.callback.onSet(Connection.this, response, System.nanoTime() - handler.startTime);
+                handler.callback.onSet(Connection.this, response, System.nanoTime() - handler.startTime, handler.retryCount);
 
                 // If we happen to be closed and we're the last outstanding request, we need to terminate the connection
                 // (note: this is racy as the signaling can be called more than once, but that's not a problem)
@@ -681,7 +680,7 @@ class Connection {
             {
                 ResponseHandler handler = iter.next();
                 handler.cancelTimeout();
-                handler.callback.onException(Connection.this, ce, System.nanoTime() - handler.startTime);
+                handler.callback.onException(Connection.this, ce, System.nanoTime() - handler.startTime, handler.retryCount);
                 iter.remove();
             }
         }
@@ -749,18 +748,24 @@ class Connection {
         }
 
         @Override
-        public void onSet(Connection connection, Message.Response response, ExecutionInfo info, Statement statement, long latency) {
-            onSet(connection, response, latency);
+        public int retryCount() {
+            // This is ignored, as there is no retry logic in this class
+            return 0;
         }
 
         @Override
-        public void onSet(Connection connection, Message.Response response, long latency) {
+        public void onSet(Connection connection, Message.Response response, ExecutionInfo info, Statement statement, long latency) {
+            onSet(connection, response, latency, 0);
+        }
+
+        @Override
+        public void onSet(Connection connection, Message.Response response, long latency, int retryCount) {
             this.address = connection.address;
             super.set(response);
         }
 
         @Override
-        public void onException(Connection connection, Exception exception, long latency) {
+        public void onException(Connection connection, Exception exception, long latency, int retryCount) {
             // If all nodes are down, we will get a null connection here. This is fine, if we have
             // an exception, consumers shouldn't assume the address is not null.
             if (connection != null)
@@ -769,10 +774,11 @@ class Connection {
         }
 
         @Override
-        public void onTimeout(Connection connection, long latency) {
+        public boolean onTimeout(Connection connection, long latency, int retryCount) {
             assert connection != null; // We always timeout on a specific connection, so this shouldn't be null
             this.address = connection.address;
             super.setException(new ConnectionException(connection.address, "Operation timed out"));
+            return true;
         }
 
         public InetSocketAddress getAddress() {
@@ -782,9 +788,10 @@ class Connection {
 
     interface ResponseCallback {
         public Message.Request request();
-        public void onSet(Connection connection, Message.Response response, long latency);
-        public void onException(Connection connection, Exception exception, long latency);
-        public void onTimeout(Connection connection, long latency);
+        public int retryCount();
+        public void onSet(Connection connection, Message.Response response, long latency, int retryCount);
+        public void onException(Connection connection, Exception exception, long latency, int retryCount);
+        public boolean onTimeout(Connection connection, long latency, int retryCount);
     }
 
     static class ResponseHandler {
@@ -792,6 +799,7 @@ class Connection {
         public final Connection connection;
         public final int streamId;
         public final ResponseCallback callback;
+        public final int retryCount;
 
         private final Timeout timeout;
         private final long startTime;
@@ -800,6 +808,7 @@ class Connection {
             this.connection = connection;
             this.streamId = connection.dispatcher.streamIdHandler.next();
             this.callback = callback;
+            this.retryCount = callback.retryCount();
 
             long timeoutMs = connection.factory.getReadTimeoutMillis();
             this.timeout = timeoutMs <= 0 ? null : connection.factory.timer.newTimeout(onTimeoutTask(), timeoutMs, TimeUnit.MILLISECONDS);
@@ -826,7 +835,8 @@ class Connection {
             return new TimerTask() {
                 @Override
                 public void run(Timeout timeout) {
-                    callback.onTimeout(connection, System.nanoTime() - startTime);
+                    if (callback.onTimeout(connection, System.nanoTime() - startTime, retryCount))
+                        cancelHandler();
                 }
             };
         }